requirements/dev.txt
-r base.txt
pytest-django
coverage
selenium
faker
factory_boy